Split board storage responsibilities
