+++ src/features/board/board-sets/board-set-library.tsx
@@ -7,13 +7,13 @@ import ListItemButton from "@mui/material/ListItemButton";
 import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
 import { m } from "@paraglide/messages.js";
-import { EmptyState } from "@shared/ui/empty-state";
-import { LoadingState } from "@shared/ui/loading-state";
 import { useTranslate } from "@shared/language/use-translate";
 import { useSnackbar } from "@shared/snackbar/use-snackbar";
+import { EmptyState } from "@shared/ui/empty-state";
+import { LoadingState } from "@shared/ui/loading-state";
 import { useState } from "react";
 import { useImportBoardFiles } from "../import/use-import-board-files";
-import type { BoardSetRecord } from "../storage/boards-db";
+import type { BoardSetRecord } from "../storage/board-set-storage";
 import { BoardSetDeleteDialog } from "./board-set-delete-dialog";
 import { BoardSetInfoDialog } from "./board-set-info-dialog";
 import { BoardSetList } from "./board-set-list";
