+++ src/features/board/board-sets/board-set-list.tsx
@@ -15,7 +15,7 @@ import Tooltip from "@mui/material/Tooltip";
 import { m } from "@paraglide/messages.js";
 import { useTranslate } from "@shared/language/use-translate";
 import { useState } from "react";
-import type { BoardSetRecord } from "../storage/boards-db";
+import type { BoardSetRecord } from "../storage/board-set-storage";
 
 interface BoardSetListProps {
   boardSets: BoardSetRecord[];
