+++ src/features/board/navigation/use-boards-in-set.ts
@@ -1,6 +1,6 @@
 import { useLatestAsync } from "@shared/hooks/use-latest-async";
 import { useLanguage } from "@shared/language/use-language";
-import { listBoards } from "../storage/boards-db";
+import { listBoards } from "../storage/board-content-storage";
 import { findTranslations } from "../translation/board-strings";
 
 interface UseBoardsInSetOptions {
