+++ src/features/board/storage/board-hydration.test.ts
@@ -1,10 +1,11 @@
+import { assertDefined } from "@shared/testing/assert-defined";
 import type { OBFBoard } from "@shayc/open-board-format";
 import { beforeEach, describe, expect, test } from "vitest";
-import { assertDefined } from "@shared/testing/assert-defined";
 import { refreshBoardSets } from "../board-sets/board-sets-store";
-import { hydrateBoard, type HydratedBoard } from "./board-hydration";
-import { BoardNotFoundError, createBoardSet } from "./boards-db";
 import { resetBoardsDB } from "../testing";
+import { BoardNotFoundError } from "./board-content-storage";
+import { hydrateBoard, type HydratedBoard } from "./board-hydration";
+import { createBoardSet } from "./board-set-storage";
 
 const SET_ID = "loader-test-set";
 const BOARD_ID = "loader-test-board";
