+++ src/features/board/testing/index.ts
@@ -1,2 +1,7 @@
-export { makeOBFBoard, resetBoardsDB, seedBoardSets } from "./db";
+export {
+  countStoredBoardContent,
+  makeOBFBoard,
+  resetBoardsDB,
+  seedBoardSets,
+} from "./db";
 export { loadFixtureFile, TEST_IMAGE_SRC } from "./fixtures";
